+++ FinBud-Frontend/src/app/quiz/page.tsx
@@ -6,18 +6,18 @@ import {
   Node,
   NodeType,
   updateAnswers,
-} from "./../../quiz-logic/quiz-functions";
+} from "../../quiz-logic/quiz-functions";
 import { FaAngleLeft, FaAngleRight } from "react-icons/fa";
 
 import Image from "next/image";
-import Button from "@/components/Button";
-import Textbox from "@/components/Textbox";
+import Button from "@/components/Back-Button-Component/Button";
+import Textbox from "@/components/Textbox-Component/Textbox";
 import { IoIosArrowRoundBack } from "react-icons/io";
-import CalculatorComponent from "../../components/Calculator Component/CalculatorComponent";
+import CalculatorComponent from "../../components/Calculator-Component/CalculatorComponent";
 import React from "react";
 
 import { useRouter } from "next/navigation";
-import CalculatorButton from "../../components/Calculator Component/CalculatorButton";
+import CalculatorButton from "../../components/Calculator-Component/CalculatorButton";
 
 export default function QuizPage() {
   const router = useRouter();
