+++ FinBud-Frontend/src/components/Calculator-Component/CalculatorComponent.tsx
