+++ FinBud-Frontend/src/components/Calculator-Component/Reference.tsx
