--- FinBud-Frontend/src/components/answer.tsx
@@ -1,15 +1,0 @@
-import React from "react";
-
-interface AnswerProps {
-  result: string;
-}
-
-const Answer: React.FC<AnswerProps> = ({ result }) => {
-  return (
-    <div>
-      <h2>{result}</h2>
-    </div>
-  );
-};
-
-export default Answer;
